+++ test/next/CHANGELOG.md
@@ -1,5 +1,19 @@
 # vercel-storage-integration-test-suite
 
+## 0.1.8
+
+### Patch Changes
+
+- Updated dependencies [97a3d06]
+  - @vercel/edge-config@0.2.1
+
+## 0.1.7
+
+### Patch Changes
+
+- Updated dependencies [7944205]
+  - @vercel/edge-config@0.2.0
+
 ## 0.1.6
 
 ### Patch Changes